// Client/src/app/interfaces/agent.ts
export interface Agent {
    MATRICULE: string;
    FONCTION_AG: string;
    MAIL_AG: string;
    NOM_AG: string;
    NOM_UTIL_AG: string;
    TYPE_AG: string;
    PRENOM_AG: string;
    ADRESSE_AG: string;
    TEL_AG: string;
    PASSWORD: string;
    PHOTO?: string;
    GENRE: string;
    ACTIVATION: string;
    CODE_DIVISION: string;
}